restructures and documentation of code
