+++ index.html
@@ -21,5 +21,6 @@
 </head>
 <body>
     <div id='app'></div> <!-- div#app is where React App is rendered with ReactDOM.render in app.js-->
+    <script src="/dist/bundle.js"></script> <!-- /dist/bundle.js is our compiled JS from webpack-->
 </body>
 </html>
